% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commarobust.R
\name{commarobust}
\alias{commarobust}
\title{Title}
\usage{
commarobust(fit, cluster = NULL, type = "HC2")
}
\arguments{
\item{fit}{A model object, typically created by lm, glm, or many other model fitting functions.}

\item{cluster}{An optional numeric vector of length N that describes the clusters that units are in. May not include NAs. Is NULL by default. If specified, cluster-robust standard errors will be estimated.}

\item{type}{A string indicating the type of heterskedasticty-robust standard errors to be estimated. Is ignored if cluster is specificed. Is "HC2" by default, because these are equivalent to Neyman standard errors (See Aronow and Samii).}
}
\value{
A matrix of coefficients, standard errors, t-statistics, and p-values
}
\description{
Title
}
\examples{

library(randomizr)

Y <- rnorm(100)
Z <- complete_ra(100)
fit <- lm(Y ~ Z)
commarobust(fit)

# Clustered
Y <- rnorm(100)
clust <- rep(letters[1:10], 10)
Z <- cluster_ra(clust)

fit <- lm(Y ~ Z)
commarobust(fit, cluster = clust)

}

