% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commarobust.R
\name{makerobustpslist}
\alias{makerobustpslist}
\title{Prepare ps for Stargazer}
\usage{
makerobustpslist(..., type = "HC2")
}
\arguments{
\item{...}{a series of model fits, separated by commas}

\item{type}{A string indicating the type of heterskedasticty-robust standard errors to be estimated. Is ignored if clust_var is specificed. Is "HC2" by default, because these are equivalent to Neyman standard errors (See Aronow and Samii).}
}
\value{
A list of vectors of robust standard errors.
}
\description{
This function will prepare p-values for stargazer based on the robust standard errors. Stargazer will do this automatically, so the main purpose of this function is to set the p-value of the intercept to 1. In experimental research, the intercept is often the sample mean for the control group and a test of the null hypothesis that it is equal to do is silly to report.
}
\examples{
library(randomizr)
library(stargazer)

Y_1 <- rnorm(100)
Z_1 <- complete_ra(100)
Y_2 <- rnorm(100)
Z_2 <- complete_ra(100)

fit_1 <- lm(Y_1 ~ Z_1)
fit_2 <- lm(Y_2 ~ Z_2)

stargazer(fit_1, fit_2,
          se = makerobustseslist(fit_1, fit_2),
          p = makerobustpslist(fit_1, fit_2))

}

