% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commarobust.R
\name{makerobustseslist}
\alias{makerobustseslist}
\title{Prepare Ses for Stargazer}
\usage{
makerobustseslist(..., type = "HC2")
}
\arguments{
\item{...}{a series of model fits, separated by commas}

\item{type}{A string indicating the type of heterskedasticty-robust standard errors to be estimated. Is ignored if clust_var is specificed. Is "HC2" by default, because these are equivalent to Neyman standard errors (See Aronow and Samii).}
}
\value{
A list of vectors of robust standard errors.
}
\description{
This function will prepare heteroskedasticity-robust standard errors for stargazer. Currently, it does not handle cluster-robust standard errors because there is no nice way to pass model-specific vectors of clusters.
}
\examples{
library(randomizr)
library(stargazer)

Z_1 <- complete_ra(100)
Y_1 <- 10 + 5*Z_1 + rnorm(100)
Z_2 <- complete_ra(100)
Y_2 <- 10 + 2*Z_2 + rnorm(100)

fit_1 <- lm(Y_1 ~ Z_1)
fit_2 <- lm(Y_2 ~ Z_2)

stargazer(fit_1, fit_2,
          se = makerobustseslist(fit_1, fit_2),
          p = makerobustpslist(fit_1, fit_2))

}
