% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commarobust.R
\name{se_mean}
\alias{se_mean}
\title{Obtain the standard error of the mean}
\usage{
se_mean(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
a scalar equal to the standard error of the mean, computed as sd(x)/(sqrt(n)).
}
\description{
Obtain the standard error of the mean
}
\examples{

x <- rnorm(100)
mean(x)
se_mean(x)


}

